[rename]파일 위치 이동 (#41)
디렉토리 구조를 uilt.sort로 이동
