+++ src/main/java/org/kakaoshare/backend/domain/product/repository/query/ProductRepositoryCustomImpl.java
@@ -6,8 +6,8 @@ import com.querydsl.core.types.dsl.BooleanExpression;
 import com.querydsl.jpa.JPAExpressions;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import lombok.RequiredArgsConstructor;
-import org.kakaoshare.backend.common.util.SortUtil;
-import org.kakaoshare.backend.common.util.SortableRepository;
+import org.kakaoshare.backend.common.util.sort.SortUtil;
+import org.kakaoshare.backend.common.util.sort.SortableRepository;
 import org.kakaoshare.backend.domain.product.dto.DescriptionResponse;
 import org.kakaoshare.backend.domain.product.dto.DetailResponse;
 import org.kakaoshare.backend.domain.product.entity.query.QSimpleProductDto;
